cleaned workspace and making progress on update
need to make self_destruct module
